[0.19.x] Experimental Pluggable Query Handler
This is an experimental feature for composer and as such is not documented.
